+++ sdk/keyvault/azure-security-keyvault-keys/test/ut/key_client_base_test.hpp
@@ -12,7 +12,7 @@
 #include <azure/core/context.hpp>
 #include <azure/core/uuid.hpp>
 #include <azure/identity/client_secret_credential.hpp>
-#include <azure/keyvault/key_vault.hpp>
+#include <azure/keyvault/key_vault_keys.hpp>
 
 #include <cstdio>
 #include <iostream>