+++ sdk/keyvault/azure-security-keyvault-keys/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features
 
 - Added support for importing and deserializing EC and OCT keys.
+- Added cryptography client.
 - Added `CreateFromResumeToken()` to `DeletedKeyOperation` and `RecoverKeyOperation`.
 
 ### Breaking Changes